pyproject.toml
[project]
name = "univr-fellowships-crawler"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4==4.12.3",
    "ipykernel>=7.2.0",
    "lxml==5.3.0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "supabase==2.10.0",
]
